refactor: emailValidation to email
